requirements.txt
Django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0 